Add: Introduce McpClientApp and McpServerApp classes for improved MCP client and server functionality. Change: Refactor create_app function to utilize new classes. Update: Modify imports in utilities.py to align with new class structure.
